refactor(cli): split render/index.ts by command domain
Pure moves, no output change (golden suite byte-identical):
  family.ts  — FAMILY_RENDER hook table + renderFamily
  wallet.ts  — wallet receipts + list tree view
  account.ts — account/token queries + history/portfolio rows
  tx.ts      — signing receipts + tx status/info
  misc.ts    — config/networks/contract/message/block
index.ts stays the barrel: reassembles TextFormatters (spread) and
keeps renderGenericText, so no importer changes. methodName moves to
scalars.ts (shared by tx receipts and contract call).
